feat: tag statistics 캐싱 적용(write through)
